backend/requirements.txt
fastapi>=0.95.1
uvicorn>=0.22.0
sqlalchemy>=2.0.9
pydantic>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
alembic==1.12.1
pydantic[email]
pydantic-settings>=2.0.0
pandas>=2.0.0
xlsxwriter>=3.1.0
sacrebleu>=2.3.1
pyyaml>=6.0
python-crontab>=2.7.0
# Uncomment the line below if using COMET score evaluation
# unbabel-comet>=2.0.0